transfer category selection to MemoryUtils
