+++ app/src/main/java/vn/edu/rmit/circlelink/MemoryUtils.java
@@ -1,6 +1,9 @@
 package vn.edu.rmit.circlelink;
 
+import android.content.Context;
 import android.util.Log;
+
+import androidx.appcompat.app.AlertDialog;
 
 import java.text.SimpleDateFormat;
 import java.time.LocalDate;
@@ -16,6 +19,7 @@ import vn.edu.rmit.circlelink.model.Memory;
 public class MemoryUtils {
 
     public static ArrayList<Memory> currentMemories = new ArrayList<>();
+    private static final String[] categories = {"Travel", "Birthdays", "Hangouts", "Celebrations", "Holidays"};
 
     public static LinkedHashMap<String, ArrayList<Memory>> groupAndSortMemoriesByMonth() {
 
@@ -32,6 +36,40 @@ public class MemoryUtils {
     public static String formatLocalDateToMonthYear(LocalDate date) {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MMMM yyyy");
         return date.format(formatter);
+    }
+
+    public static void showCategorySelectionDialog(Context context, CategorySelectionCallback callback) {
+        // Define the categories
+        final String[] categories = {"Travel", "Birthdays", "Hangouts", "Celebrations", "Holidays"};
+        final String[] selectedCategory = {null}; // Temporary holder for the selected category
+
+        AlertDialog.Builder builder = new AlertDialog.Builder(context);
+        builder.setTitle("Select a Category (optional)");
+
+        // Use single-choice items (radio button)
+        builder.setSingleChoiceItems(categories, -1, (dialog, which) -> {
+            selectedCategory[0] = categories[which];
+        });
+
+        builder.setPositiveButton("Confirm", (dialog, which) -> {
+            if (selectedCategory[0] != null) {
+                callback.onCategorySelected(selectedCategory[0]);
+            } else {
+                callback.onCategorySelected("none"); // Default to "none" if nothing is selected
+            }
+            dialog.dismiss();
+        });
+
+        builder.setNegativeButton("Skip", (dialog, which) -> {
+            callback.onCategorySelected("none"); // Default to "none" on cancel
+            dialog.dismiss();
+        });
+
+        builder.create().show();
+    }
+
+    public interface CategorySelectionCallback {
+        void onCategorySelected(String category);
     }
 
 }
